backend/requirements/production.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements/production.txt requirements/production.in
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via jsonschema
cachecontrol==0.12.11
    # via firebase-admin
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
cryptography==38.0.3
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth[with_social]==2.2.5
    # via -r requirements/production.in
django==4.0.4
    # via
    #   -r requirements/production.in
    #   dj-rest-auth
    #   django-allauth
    #   django-model-utils
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.50.0
    # via dj-rest-auth
django-environ==0.9.0
    # via -r requirements/production.in
django-model-utils==4.3.1
    # via -r requirements/production.in
djangorestframework==3.13.1
    # via
    #   -r requirements/production.in
    #   dj-rest-auth
    #   drf-spectacular
drf-spectacular==0.24.2
    # via -r requirements/production.in
firebase-admin==6.0.1
    # via -r requirements/production.in
google-api-core[grpc]==2.10.2
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.66.0
    # via firebase-admin
google-auth==2.14.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.2
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.7.2
    # via firebase-admin
google-cloud-storage==2.6.0
    # via firebase-admin
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.0
    # via google-cloud-storage
googleapis-common-protos==1.57.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.50.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.50.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements/production.in
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.17.0
    # via drf-spectacular
msgpack==1.0.4
    # via cachecontrol
oauthlib==3.2.2
    # via requests-oauthlib
proto-plus==1.22.1
    # via google-cloud-firestore
protobuf==4.21.9
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2==2.9.5
    # via -r requirements/production.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via
    #   django-allauth
    #   firebase-admin
pyparsing==3.0.9
    # via httplib2
pyrsistent==0.19.2
    # via jsonschema
python3-openid==3.2.0
    # via django-allauth
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
requests==2.28.1
    # via
    #   cachecontrol
    #   django-allauth
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via
    #   drf-spectacular
    #   google-api-python-client
urllib3==1.26.12
    # via requests
whitenoise==6.2.0
    # via -r requirements/production.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
